Moved wifi files into seperate folder
