+++ firmware/main/CloudLink/CloudLinkTask.h
@@ -11,7 +11,7 @@
 #include "Events/WifiSettingsEvent.h"
 #include "Events/WifiStatusEvent.h"
 #include "Events/SensorDataEvent.h"
-#include "Wifi.h"
+#include "Wifi/Wifi.h"
 #include "CloudLinkSenderIfc.h"
 #include "ConfigurationService/ConfigurationService.h"
 #include "MqttService/MqttService.h"
